Trait odysseus::bc::traits::Parent

pub trait Parent<Child> where
    Child: Element, {
    fn is_finished(&self, id: Child::Id) -> bool;

    fn get(&self, id: Child::Id) -> Option<&Child>;

    fn create_builder(&mut self, id: Child::Id) -> Child::Builder;

    fn finish(&mut self, id: Child::Id) -> &Child;
}

A bytecode parent, i.e. a structure that contains one or more types of bytecode elements.

Required methods

fn is_finished(&self, id: Child::Id) -> bool

Check if a bytecode element is finished. This function should return false if the ID was not present within the parent. This function is called by Entry::is_finished.

fn get(&self, id: Child::Id) -> Option<&Child>

Get a reference to a bytecode element. This function is called by Entry::get.

fn create_builder(&mut self, id: Child::Id) -> Child::Builder

Create a builder from the specified bytecode element ID.

Panics

This function may panic if the ID was not present within the parent, depending on the implementor.

fn finish(&mut self, id: Child::Id) -> &Child

Finish the bytecode element with the specified ID.

Panics

This function may panic if the ID was not present within the parent, depending on the implementor.

Loading content...